Refactor packages to ch.epfl.balelecbud
