+++ app/src/main/java/ch/epfl/balelecbud/GreetingActivity.java
@@ -1,15 +1,10 @@
-package ch.epfl.bootcamp;
+package ch.epfl.balelecbud;
 
 import android.content.Intent;
 import android.os.Bundle;
 
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
-
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 
-import android.view.View;
 import android.widget.TextView;
 
 public class GreetingActivity extends AppCompatActivity {
